rename localparser to gen
